+++ app/src/screens/passport/PassportNFCTroubleScreen.tsx
@@ -13,10 +13,9 @@ import { Caption } from '@/components/typography/Caption';
 import { useFeedbackAutoHide } from '@/hooks/useFeedbackAutoHide';
 import useHapticNavigation from '@/hooks/useHapticNavigation';
 import SimpleScrolledTitleLayout from '@/layouts/SimpleScrolledTitleLayout';
-import analytics from '@/utils/analytics';
+import analytics, { flushAllAnalytics } from '@/utils/analytics';
 import { slate500 } from '@/utils/colors';
 import { sendFeedbackEmail } from '@/utils/email';
-import { flushMixpanelEvents } from '@/utils/nfcScanner';
 
 const { flush: flushAnalytics } = analytics();
 
@@ -56,8 +55,7 @@ const PassportNFCTrouble: React.FC = () => {
 
   // error screen, flush analytics
   useEffect(() => {
-    flushAnalytics();
-    flushMixpanelEvents();
+    flushAllAnalytics();
   }, []);
 
   // 5-taps with a single finger
