+++ app/src/screens/prove/QRCodeTroubleScreen.tsx
@@ -9,9 +9,8 @@ import Tips from '@/components/Tips';
 import { Caption } from '@/components/typography/Caption';
 import useHapticNavigation from '@/hooks/useHapticNavigation';
 import SimpleScrolledTitleLayout from '@/layouts/SimpleScrolledTitleLayout';
-import analytics from '@/utils/analytics';
+import analytics, { flushAllAnalytics } from '@/utils/analytics';
 import { slate500 } from '@/utils/colors';
-import { flushMixpanelEvents } from '@/utils/nfcScanner';
 
 const { flush: flushAnalytics } = analytics();
 
@@ -50,8 +49,7 @@ const QRCodeTrouble: React.FC = () => {
 
   // error screen, flush analytics
   useEffect(() => {
-    flushAnalytics();
-    flushMixpanelEvents();
+    flushAllAnalytics();
   }, []);
 
   return (
