+++ app/src/utils/nfcScanner.ts
@@ -2,15 +2,13 @@
 // SPDX-License-Identifier: BUSL-1.1
 // NOTE: Converts to Apache-2.0 on 2029-06-11 per LICENSE.
 
-/* eslint-disable @typescript-eslint/no-unused-vars */
 import { Buffer } from 'buffer';
-import type { AppStateStatus } from 'react-native';
 import { NativeModules, Platform } from 'react-native';
 import { reset, scan as scanDocument } from 'react-native-passport-reader';
-import { ENABLE_DEBUG_LOGS, MIXPANEL_NFC_PROJECT_TOKEN } from '@env';
-import NetInfo from '@react-native-community/netinfo';
 
 import type { PassportData } from '@selfxyz/common/types';
+
+import { configureNfcAnalytics } from '@/utils/analytics';
 
 interface AndroidScanResponse {
   mrz: string;
@@ -38,72 +36,6 @@ interface Inputs {
   extendedMode?: boolean;
   usePacePolling?: boolean;
 }
-
-// --- Mixpanel flush strategy ---
-let mixpanelConfigured = false;
-let eventCount = 0;
-let flushTimer: ReturnType<typeof setInterval> | null = null;
-let isConnected = true;
-const eventQueue: Array<{
-  name: string;
-  properties?: Record<string, unknown>;
-}> = [];
-
-export const configureNfcAnalytics = () => {
-  if (!MIXPANEL_NFC_PROJECT_TOKEN || mixpanelConfigured) return;
-  const enableDebugLogs = JSON.parse(String(ENABLE_DEBUG_LOGS));
-  NativeModules.PassportReader.configure(
-    MIXPANEL_NFC_PROJECT_TOKEN,
-    enableDebugLogs,
-    {
-      flushInterval: 30,
-      flushCount: 5,
-      flushOnBackground: true,
-      flushOnForeground: true,
-      flushOnNetworkChange: true,
-    },
-  );
-  setupFlushPolicies();
-  mixpanelConfigured = true;
-};
-
-const setupFlushPolicies = () => {
-  if (flushTimer) return;
-  flushTimer = setInterval(flushMixpanelEvents, 30000);
-
-  AppState.addEventListener('change', (state: AppStateStatus) => {
-    if (state === 'background' || state === 'active') {
-      flushMixpanelEvents();
-    }
-  });
-
-  NetInfo.addEventListener(state => {
-    isConnected = state.isConnected ?? true;
-    if (isConnected) {
-      flushMixpanelEvents();
-    }
-  });
-};
-
-export const flushMixpanelEvents = () => {
-  if (!MIXPANEL_NFC_PROJECT_TOKEN) return;
-  try {
-    if (__DEV__) console.log('[Mixpanel] flush');
-    // Send any queued events before flushing
-    while (eventQueue.length > 0) {
-      const evt = eventQueue.shift()!;
-      NativeModules.PassportReader?.trackEvent?.(evt.name, evt.properties);
-    }
-    NativeModules.PassportReader?.flush?.();
-    eventCount = 0;
-  } catch (err) {
-    if (__DEV__) console.warn('Mixpanel flush failed', err);
-    // re-queue on failure
-    if (typeof err !== 'undefined') {
-      // no-op, events are already queued if failure happened before flush
-    }
-  }
-};
 
 export const parseScanResponse = (response: unknown) => {
   return Platform.OS === 'android'
@@ -142,29 +74,6 @@ const scanIOS = async (inputs: Inputs) => {
     inputs.extendedMode ?? false,
     inputs.usePacePolling ?? false,
   );
-};
-
-export const trackNfcEvent = (
-  name: string,
-  properties?: Record<string, unknown>,
-) => {
-  if (!MIXPANEL_NFC_PROJECT_TOKEN) return;
-  if (!mixpanelConfigured) configureNfcAnalytics();
-
-  if (!isConnected) {
-    eventQueue.push({ name, properties });
-    return;
-  }
-
-  try {
-    NativeModules.PassportReader?.trackEvent?.(name, properties);
-    eventCount++;
-    if (eventCount >= 5) {
-      flushMixpanelEvents();
-    }
-  } catch (err) {
-    eventQueue.push({ name, properties });
-  }
 };
 
 const handleResponseIOS = (response: unknown) => {
